// src/app/modules/support/models/question-type.model.ts
export interface IQuestionTypeModel {
  updateDate: string;
  createUserId: string;
  code: string;
  typeName: string;
  orderBy: string;
  searchKeyword: string;
  nameRu: string;
  parentNameEn: string;
  typeNameRu: string;
  parentCode: string;
  id: string;
  nameAz: string;
  createDate: string;
  parentNameAz: string;
  typeNameAz: string;
  parentTypeId: string;
  updateUserId: string;
  hiddenStatus: string;
  active: string;
  nameEn: string;
  parentId: string;
  parentName: string;
  typeNameEn: string;
  parentNameRu: string;
  name: string;
  typeId: string;
}